rename field map to field schema, add to remaining relevant methods
This adds optional field mapping via a schema generic to remaining
relevant methods - that is, any method which returns the same record as
the record calling the method, or any method that creates or loads a
record. Sublist and subrecord methods are not effected, except those
that return the parent record.