N/query: Add bigint as a possible query result value
It is possible for a value returned from `ResultSet.asMappedResults()`
to return a `bigint` instead of a `number` for integer number fields
with high enough values stored. This is significant, and undocumented by
NetSuite. I encountered this because `JSON.stringify()` doesn't know how
to serialize BigInt values by default, which will cause an uncatchable
and seemingly-impossible UNEXPECTED_ERROR when query result values are
returned as object properties in a RESTlet endpoint, after the endpoint
function has already successfully returned.